nh-248
SVN: trunk@1376

+++ src/NHibernate/Type/DecimalType.cs
@@ -76,7 +76,7 @@ namespace NHibernate.Type
 		/// <returns></returns>
 		public object StringToObject( string xml )
 		{
-			return long.Parse( xml );
+			return Decimal.Parse( xml );
 		}
 
 		/// <summary>
